Gripper: Add clear_calibration, sequence, acknowledgement
Adds command sequencing and acknowledgement checks to the gripper
interface communication with the (on robot) gripper plugin. Also
adds the clear_calibration command to the gripper.py interface.

* Checking for command acknowledgement fixes many issues related
  to False Positives with gripper command success checks.

* Also helps significantly with dealing with errors in reboots,
  resets, and calibrations.

* Clear calibration command should be used for redoing calibrated
  grippers without rebooting.

* Blocking commands now first do a wait_for() testing for command
  acknowledgement and then do a wait_for() for the provided test().

NOTE: Due to a minor known bug in certain gripper state-machine
commands that clears the sequence fields, certain command
acknowledge checks also check for a sequence=0. This works on the
assumption that the sequence is always incremented.