Removed camera close before reload open call.
Recent changes in Indigo Baxter require that only two cameras ever be
powered on at the same. Cameras that are powered off cannot be listed by
the ROS service list camera call. A fix is to power down a camera being
closed, and power up both other cameras -wait 5 seconds- then enumerate
all cameras. This will allow the other two cameras to be listed by a
subsequent call to ROS service list cameras. A side effect of this
workaround is that no camera can be closed and opened again immediately
without first closing another camera. The SDK currently closes and opens
the cameras by default and to reload camera settings, which won't work
with this fix. As it turns out, cameras that are opened and then
reopened works, and will give the same "settings reload" functionality,
so there is no need to close a camera before opening it. Therefore
the unnecessary camera close was removed from camera.py to
prevent toggling power to two separate cameras before opening the
desired camera.