Fixed Bezier lib to prevent errors on 2 points
This commit partially resolves baxter_interface issue (#53 and #54),
where the JTAS would error out if too few points were passed into it.

There was a bug in the Bezier library which did not properly check for
two points, which this resolves. We can safely remove the additional
two point check in the joint trajectory action server based on this.

Trac ticket #11669