JTAS copies initial vels & accels if one traj pt
The if only one trajectory point is provided to the Joint Trajectory
Action Server, it technically does not count as a valid trajectory.
However, we attempt to make do by using the current position as the
first point in the trajectory. The second (and only point) is then
the provided trajectory pose.

The JTAS zeros out both velocities and accelerations for the final
position, since it assumes each trajectory ends with a stopped motion.
To prevent loss of data, in the case of only 1 trajectory point being
provided, this commit copies out the only point's velocities and
accelerations into the created initial point consisting of the current
joint positions. Having only one trajectory point is an edge case,
but one we would like to handle in a logical manner.